setup.py: added setup.py and created 'package'
this commit moves all of the original files into the following directories

- `bin/` for command-line executables
- `bruco/` for the core python library
- `share/` for distributed txt/configuration files
