+++ bruco/__init__.py
@@ -1,0 +1,4 @@
+# Brute force coherence (Gabriele Vajente, 2015-06-09)
+
+"""Brute-force coherence analysis of LIGO data
+"""
