+++ share/bruco_excluded_channels.txt
